PAY-261 #190507 Fix setting a restart interval for goroutines returning product keys from the reserve.

+++ internal/app.go
@@ -324,7 +324,7 @@ func (app *Application) KeyDaemonStart() {
 	zap.S().Infof("Key daemon started", zap.Int("RestartInterval", app.cfg.KeyDaemonRestartInterval))
 
 	go func() {
-		interval := time.Duration(int64(app.cfg.KeyDaemonRestartInterval) * int64(time.Second))
+		interval := time.Duration(app.cfg.KeyDaemonRestartInterval) * time.Second
 		shutdown := make(chan os.Signal, 1)
 		signal.Notify(shutdown, syscall.SIGTERM, syscall.SIGINT, syscall.SIGQUIT)
