+++ internal/config/config.go
@@ -73,7 +73,7 @@ type Config struct {
 	PaysuperDocumentSignerEmail string `envconfig:"PAYSUPER_DOCUMENT_SIGNER_EMAIL" required:"true"`
 	PaysuperDocumentSignerName  string `envconfig:"PAYSUPER_DOCUMENT_SIGNER_NAME" required:"true"`
 
-	KeyDaemonRestartInterval int `envconfig:"KEY_DAEMON_RESTART_INTERVAL" default:"60"`
+	KeyDaemonRestartInterval int64 `envconfig:"KEY_DAEMON_RESTART_INTERVAL" default:"60"`
 
 	*PaymentSystemConfig
 	*CustomerTokenConfig
